mcp-client: introduce ScopeStepUpMcpOAuth2ClientManager for cases with DCR disabled
